requirements.txt
pillow~=10.2.0
ipython~=8.18.0
langdetect~=1.0.9
googletrans~=4.0.0rc1
streamlit~=1.29.0
Flask~=3.0.0
validators~=0.22.0
gunicorn==21.2.0
google-generativeai==0.3.2